Force a passive scan on explicit scan requests
libwpa_client on FreeBSD doesn't perform a scan and falls back to using
the last scan cache when the device is connected to a network. To ensure
a scan is performed on `$ wutil scan` command and using the scan keybinding on
`wutui`, we briefly set passive_scan=1 on wpa_supplicant which ensures
an actual scan is performed.